Update the types of the arguments for General Commissioning cluster's ArmFailSafe command to match spec

+++ rs-matter/src/data_model/sdm/general_commissioning.rs
@@ -117,8 +117,8 @@ pub const CLUSTER: Cluster<'static> = Cluster {
 
 #[derive(FromTLV, ToTLV)]
 struct FailSafeParams {
-    expiry_len: u8,
-    bread_crumb: u8,
+    expiry_len: u16,
+    bread_crumb: u64,
 }
 
 #[derive(ToTLV)]
